moving deserializer into serialization module
