+++ src/pydase/client/client.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, TypedDict
 import socketio  # type: ignore
 
 from pydase.client.proxy_class_factory import ProxyClassFactory
-from pydase.utils.deserializer import loads
+from pydase.utils.serialization.deserializer import loads
 from pydase.utils.serializer import SerializedObject
 
 if TYPE_CHECKING:
